+++ docs/whats-new.rst
@@ -4,9 +4,11 @@ What's New
 ==========
 
 v0.1.8 (X XXX. 2021)
------------------------
+--------------------
 
 **Features and front-end API**
+
+- Add Speed of sound computation to xarray accessor teos10 method. (:pr:`90`) by `G. Maze <http://www.github.com/gmaze>`_.
 
 - Code spell fixes (:pr:`89`) by `K. Schwehr <https://github.com/schwehr>`_.